requirements.txt
# Web & API
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
python-dotenv==1.0.1

# Machine Learning & Data Processing
pandas>=2.2.3,<3.0
numpy>=2.1.2,<3.0
scikit-learn>=1.5.2,<2.0
joblib==1.4.2
matplotlib>=3.9.2,<4.0
dnspython==2.8.0

# Feature Extraction
tldextract>=4.0.0,<6.0

# OpenSearch
opensearch-py>=3.1.0,<4.0.0
